pyproject.toml
[project]
name = "python-ads-client"
description = "A python client for communicating with a Beckhoff PLC via ADS"

dynamic = ["version"]

readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [
  { email = "matthew@davidson.engineering" },
  { name = "Matthew Davidson" },
]

classifiers = [
  "Development Status :: 1 - Planning",
  "Operating System :: Microsoft :: Windows",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]

dependencies = [
  "pyads>=3.4.2",
  "pyyaml>=6.0",
  "prometheus_client>=0.2.0",
  'python-json-logger>=2.0.7',
  "python-config-loader @ git+https://github.com/davidson-engineering/python-config-loader.git@v0.0.2",
  "buffered @ git+https://github.com/generalmattza/buffered.git@v1.0.1",
]

[tool.setuptools.dynamic]
version = { attr = "ads_client.__version__" }

[project.optional-dependencies]
test = ["pytest >= 7.1.1"]

# [tool.pytest.ini_options]
# log_cli = true
# log_cli_level = "CRITICAL"
# log_cli_format = "%(message)s"
# addopts = "-n 10"

# [project.urls]
# homepage = "https://example.com"
# documentation = "https://readthedocs.org"
# repository = "https://github.com"
# changelog = "https://github.com/me/spam/blob/master/CHANGELOG.md"

# [project.scripts]
# spam-cli = "spam:main_cli"

# [project.gui-scripts]
# spam-gui = "spam:main_gui"

# [project.entry-points."spam.magical"]
# tomatoes = "spam:main_tomatoes"
